Fixed spawning
Spawning works fine, but not truly as intended. Enemies spawn within a radius of ~150 meters, and the user can only see within a radius of 100 meters as intended. However, the spawn cap does not include the enemies within the radius of 100-150 meters because the user cannot see those enemies. This is because the number of enemies near the user is calculated by the number of annotations, which is based only on what the user can see.